% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_lavaan_comparison.R
\name{bootstrap_lavaan_comparison}
\alias{bootstrap_lavaan_comparison}
\title{Bootstrap Comparison for Lavaan Models}
\usage{
bootstrap_lavaan_comparison(
  model1,
  model2,
  R = 1000,
  parallel = "no",
  ncpus = 1
)
}
\arguments{
\item{model1}{First lavaan model}

\item{model2}{Second lavaan model}

\item{R}{Number of bootstrap replications (default: 1000)}

\item{parallel}{Type of parallel processing (if any)}

\item{ncpus}{Number of CPUs to use for parallel processing}
}
\value{
A list containing bootstrap results
}
\description{
Perform bootstrap-based comparison of lavaan models
}
\examples{
\dontrun{
library(lavaan)
model1 <- 'F1 =~ x1 + x2 + x3'
model2 <- 'F1 =~ x1 + x2 + x3 + x4'
fit1 <- cfa(model1, data = HolzingerSwineford1939)
fit2 <- cfa(model2, data = HolzingerSwineford1939)
boot_result <- bootstrap_lavaan_comparison(fit1, fit2, R = 100)
print(boot_result)
}
}
