% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_latent_models.R
\name{compare_latent_models}
\alias{compare_latent_models}
\title{Compare Latent Variable Models}
\usage{
compare_latent_models(
  ...,
  nested = FALSE,
  fit_measures = c("chisq", "df", "pvalue", "cfi", "tli", "rmsea", "srmr"),
  method = "default"
)
}
\arguments{
\item{...}{lavaan model objects}

\item{nested}{logical, whether models are nested}

\item{fit_measures}{character vector of fit measures to extract}

\item{method}{comparison method}
}
\value{
A latent_comparison object
}
\description{
Compare Latent Variable Models
}
