% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models_advanced_lv.R
\name{compare_models_advanced_lv}
\alias{compare_models_advanced_lv}
\title{Advanced Model Comparison with Latent Variable Support}
\usage{
compare_models_advanced_lv(
  models,
  model_names = NULL,
  model_types = NULL,
  criteria = c("AIC", "BIC", "CFI", "TLI", "RMSEA", "SRMR"),
  latent_indicators = NULL,
  bootstrap = FALSE,
  n_bootstrap = 1000
)
}
\arguments{
\item{models}{List of model objects (lm, lavaan)}

\item{model_names}{Character vector of model names}

\item{model_types}{Character vector of model types ("lm", "lavaan")}

\item{criteria}{Criteria to calculate}

\item{latent_indicators}{List of latent variable indicators (for lavaan)}

\item{bootstrap}{Logical, whether to perform bootstrapping}

\item{n_bootstrap}{Number of bootstrap replications}
}
\value{
Comparison results with latent variable support
}
\description{
Advanced Model Comparison with Latent Variable Support
}
