% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Package startup message}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{library location}

\item{pkgname}{package name}
}
\description{
Package startup message
}
