% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_latent_fit.R
\name{extract_latent_fit}
\alias{extract_latent_fit}
\title{Extract Latent Variable Fit Indices}
\usage{
extract_latent_fit(model)
}
\arguments{
\item{model}{lavaan model object}
}
\value{
Comprehensive fit indices for latent variable model
}
\description{
Extract Latent Variable Fit Indices
}
