% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_pvalues}
\alias{format_pvalues}
\title{Format p-values for display}
\usage{
format_pvalues(pvals, digits = 3)
}
\arguments{
\item{pvals}{Vector of p-values}

\item{digits}{Number of digits (default=3)}
}
\value{
Formatted p-values as character strings
}
\description{
This internal function formats p-values for display, converting very
small values (e.g., < 0.001) to appropriate string representations.
}
\examples{
\dontrun{
format_pvalues(c(0.00001, 0.0345, 0.5, NA))
}
}
\keyword{internal}
