% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelscompete4-package.R
\docType{package}
\name{modelscompete4-package}
\alias{modelscompete4}
\alias{modelscompete4-package}
\title{modelscompete4: Advanced Model Comparison for Latent Variable Models}
\description{
The modelscompete4 package provides advanced tools for comparing latent
variable models, including bootstrap comparisons, fit index extraction,
and visualization functions.
}
\section{Main functions}{

\describe{
  \item{\code{\link{compare_latent_models}}}{Compare multiple latent variable models}
  \item{\code{\link{extract_latent_parameters}}}{Extract parameters from lavaan models}
  \item{\code{\link{plot_latent_comparison}}}{Visualize model comparison results}
  \item{\code{\link{bootstrap_lavaan_comparison}}}{Bootstrap-based model comparison}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ssjerf-stack/modelscompete4}
  \item Report bugs at \url{https://github.com/ssjerf-stack/modelscompete4/issues}
}

}
\author{
\strong{Maintainer}: Jerf Yeung \email{ssjerf@gmail.com}

}
\keyword{internal}
