% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_latent_comparison.R
\name{plot_latent_comparison}
\alias{plot_latent_comparison}
\title{Plot Latent Model Comparison Results}
\usage{
plot_latent_comparison(x, type = "fit", ...)
}
\arguments{
\item{x}{An object of class 'modelscompete4' or 'latent_comparison'}

\item{type}{Type of plot: 'fit' for fit indices, 'diff' for differences}

\item{...}{Additional arguments passed to plotting functions}
}
\value{
A ggplot object (if ggplot2 and tidyr are available), otherwise NULL
}
\description{
Creates visualization of model comparison results
}
\examples{
\dontrun{
library(lavaan)
model1 <- 'F1 =~ x1 + x2 + x3'
fit1 <- cfa(model1, data = HolzingerSwineford1939)
result <- compare_latent_models(fit1)
plot_latent_comparison(result)
}
}
