% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_latent_models.R,
%   R/print_latent_comparison.R
\name{print.latent_comparison}
\alias{print.latent_comparison}
\title{Print method for latent_comparison objects}
\usage{
\method{print}{latent_comparison}(x, digits = 3, ...)

\method{print}{latent_comparison}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class 'latent_comparison'}

\item{digits}{Number of digits to display (default: 3)}

\item{...}{Additional arguments passed to print method}
}
\value{
Invisibly returns the input object
}
\description{
Prints a summary of latent model comparison results.
}
