% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.modelscompete4.R
\name{print.modelscompete4}
\alias{print.modelscompete4}
\title{Print Method for modelscompete4}
\usage{
\method{print}{modelscompete4}(x, ...)
}
\arguments{
\item{x}{modelscompete4 object}

\item{...}{Additional arguments}
}
\description{
Print Method for modelscompete4
}
