% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_method.R
\name{print.modelscompete4_advanced}
\alias{print.modelscompete4_advanced}
\title{Print Method for modelscompete4_advanced Objects}
\usage{
\method{print}{modelscompete4_advanced}(x, ...)
}
\arguments{
\item{x}{A modelscompete4_advanced object}

\item{...}{Additional arguments passed to print}
}
\description{
Print Method for modelscompete4_advanced Objects
}
